Extract CreateChatForm
